# backend/openapi/journal-baseline.openapi.json
{
  "components": {
    "schemas": {
...
      "AmbiguousOpenPositionDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "const": "AMBIGUOUS_OPEN_POSITION",
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "AmbiguousOpenPositionDetail",
        "type": "object"
      },
      "AssetCoreTypeEnum": {
        "enum": [
          "STOCK",
          "BOND",
          "FUND",
          "COMMODITY",
          "FX",
          "DERIVATIVE",
          "CRYPTO"
        ],
        "title": "AssetCoreTypeEnum",
        "type": "string"
      },
      "AssetCurrencyEnum": {
        "enum": [
          "USD",
          "HKD",
          "CNY",
          "EUR",
          "GBP"
        ],
        "title": "AssetCurrencyEnum",
        "type": "string"
      },
...
      "AssetMetadataResponse": {
        "properties": {
          "core_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetCoreTypeEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "currency": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetCurrencyEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "instrument": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instrument"
          },
          "market": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetMarketEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "name",
          "core_type",
          "market",
          "currency",
          "sector",
          "instrument"
        ],
        "title": "AssetMetadataResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PositionCreate": {
        "additionalProperties": false,
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "integer"
          },
          "asset_metadata": {
            "$ref": "#/components/schemas/PositionInstrumentIdentityCreate"
          },
          "asset_type": {
            "pattern": "^[\\t\\n\\v\\f\\r ]*(?:[Ss][Tt][Oo][Cc][Kk]|[Ff][Uu][Nn][Dd]|[Cc][Rr][Yy][Pp][Tt][Oo]|[Ee][Qq][Uu][Ii][Tt][Yy]|[Ee][Tt][Ff]|[Ss][Pp][Oo][Tt]_[Cc][Rr][Yy][Pp][Tt][Oo])[\\t\\n\\v\\f\\r ]*$",
            "title": "Asset Type",
            "type": "string",
            "x-canonical-values": [
              "STOCK",
              "FUND",
              "CRYPTO"
            ],
            "x-normalization": "ASCII_TRIM_UPPER"
          },
          "checklist_responses": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checklist Responses"
          },
          "direction": {
            "$ref": "#/components/schemas/PositionDirectionEnum"
          },
          "entry_confidence": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Confidence"
          },
          "entry_emotion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Emotion"
          },
          "entry_price": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Entry Price"
          },
          "entry_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Reason"
          },
          "entry_time": {
            "format": "date-time",
            "title": "Entry Time",
            "type": "string"
          },
          "exchange_code": {
            "pattern": "^[\\t\\n\\v\\f\\r ]*[A-Za-z0-9][A-Za-z0-9._-]{0,31}[\\t\\n\\v\\f\\r ]*$",
            "title": "Exchange Code",
            "type": "string",
            "x-normalization": "ASCII_TRIM_UPPER",
            "x-normalized-pattern": "^[A-Z0-9][A-Z0-9._-]{0,31}$"
          },
          "planned_entry_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Entry Price"
          },
          "planned_stop_loss": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Stop Loss"
          },
          "planned_take_profit": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Take Profit"
          },
          "quantity": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Quantity"
          },
          "strategy_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy Id"
          },
          "symbol": {
            "pattern": "^[\\t\\n\\v\\f\\r ]*[A-Za-z0-9][A-Za-z0-9._/-]{0,49}[\\t\\n\\v\\f\\r ]*$",
            "title": "Symbol",
            "type": "string",
            "x-normalization": "ASCII_TRIM_UPPER",
            "x-normalized-pattern": "^[A-Z0-9][A-Z0-9._/-]{0,49}$"
          }
        },
        "required": [
          "account_id",
          "symbol",
          "exchange_code",
          "asset_type",
          "direction",
          "entry_price",
          "quantity",
          "entry_time",
          "asset_metadata"
        ],
        "title": "PositionCreate",
        "type": "object"
      },
      "PositionCreateConflictResponse": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "discriminator": {
              "mapping": {
                "AMBIGUOUS_OPEN_POSITION": "#/components/schemas/AmbiguousOpenPositionDetail",
                "OPEN_POSITION_EXISTS": "#/components/schemas/OpenPositionExistsDetail"
              },
              "propertyName": "code"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/OpenPositionExistsDetail"
              },
              {
                "$ref": "#/components/schemas/AmbiguousOpenPositionDetail"
              }
            ],
            "title": "Detail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "PositionCreateConflictResponse",
        "type": "object"
      },
...
      "PositionDirectionEnum": {
        "enum": [
          "LONG",
          "SHORT"
        ],
        "title": "PositionDirectionEnum",
        "type": "string"
      },
...
      "PositionListResponse": {
        "description": "Lighter response for list view (without batches)",
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "asset_metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetMetadataResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "asset_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Type"
          },
          "average_entry_price": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Entry Price"
          },
          "batches": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/TradeBatchResponse"
            },
            "title": "Batches",
            "type": "array"
          },
          "closed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "direction": {
            "$ref": "#/components/schemas/PositionDirectionEnum"
          },
          "exchange": {
            "title": "Exchange",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "opened_at": {
            "format": "date-time",
            "title": "Opened At",
            "type": "string"
          },
          "public_id": {
            "title": "Public Id",
            "type": "string"
          },
          "realized_pnl": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Realized Pnl",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/PositionStatusEnum"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "total_quantity": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Quantity",
            "type": "string"
          },
          "truth_position_public_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Truth Position Public Id"
          }
        },
        "required": [
          "id",
          "public_id",
          "account_id",
          "symbol",
          "exchange",
          "direction",
          "status",
          "total_quantity",
          "average_entry_price",
          "realized_pnl",
          "opened_at",
          "closed_at",
          "created_at"
        ],
        "title": "PositionListResponse",
        "type": "object"
      },
      "PositionResponse": {
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "asset_metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetMetadataResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "asset_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Type"
          },
          "average_entry_price": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Entry Price"
          },
          "batches": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/TradeBatchResponse"
            },
            "title": "Batches",
            "type": "array"
          },
          "checklist_completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checklist Completed At"
          },
          "checklist_responses": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checklist Responses"
          },
          "closed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "direction": {
            "$ref": "#/components/schemas/PositionDirectionEnum"
          },
          "drift_analysis": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drift Analysis"
          },
          "exchange": {
            "title": "Exchange",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "lessons": {
            "items": {
              "type": "string"
            },
            "title": "Lessons",
            "type": "array"
          },
          "opened_at": {
            "format": "date-time",
            "title": "Opened At",
            "type": "string"
          },
          "planned_entry_price": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Entry Price"
          },
          "planned_stop_loss": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Stop Loss"
          },
          "planned_take_profit": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Take Profit"
          },
          "public_id": {
            "title": "Public Id",
            "type": "string"
          },
          "rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "realized_pnl": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Realized Pnl",
            "type": "string"
          },
          "screenshots": {
            "items": {
              "type": "string"
            },
            "title": "Screenshots",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/PositionStatusEnum"
          },
          "strategy_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy Id"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "total_quantity": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Quantity",
            "type": "string"
          },
          "trade_review": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trade Review"
          },
          "truth_position_public_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Truth Position Public Id"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "public_id",
          "user_id",
          "account_id",
          "strategy_id",
          "symbol",
          "exchange",
          "direction",
          "status",
          "total_quantity",
          "average_entry_price",
          "realized_pnl",
          "opened_at",
          "closed_at",
          "trade_review",
          "screenshots",
          "lessons",
          "rating",
          "created_at",
          "updated_at"
        ],
        "title": "PositionResponse",
        "type": "object"
      },
      "PositionStatusEnum": {
        "enum": [
          "OPEN",
          "CLOSED"
        ],
        "title": "PositionStatusEnum",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "\u4ea4\u6613\u65e5\u5fd7\u7cfb\u7edf - \u8bb0\u5f55\u3001\u5206\u6790\u548c\u590d\u76d8\u4ea4\u6613",
    "title": "Trading Noobs API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/positions": {
      "get": {
        "description": "List all positions for the current user",
        "operationId": "list_positions_api_positions_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PositionStatusEnum"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "symbol",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          },
          {
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          },
          {
            "in": "query",
            "name": "asset_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Type"
            }
          },
          {
            "in": "query",
            "name": "core_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Core Type"
            }
          },
          {
            "in": "query",
            "name": "market",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Market"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PositionListResponse"
                  },
                  "title": "Response List Positions Api Positions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Positions",
        "tags": [
          "positions"
        ]
      },
      "post": {
        "description": "Create a new position with the first entry batch",
        "operationId": "create_position_api_positions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionCreateConflictResponse"
                }
              }
            },
            "description": "A same-side lifecycle already exists; recover with ADD."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Position",
        "tags": [
          "positions"
        ]
      }
    },
...
  }
}